fix: копирование .env.sample в том случае если нет .env

+++ init.js
@@ -1,6 +1,10 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const fs = require('fs')
 
-fs.copyFileSync('.env.sample', '.env')
+
+// проверка на существование файла .env
+if (!fs.existsSync('.env')) {
+  fs.copyFileSync('.env.sample', '.env')
+}
 
 fs.mkdirSync('tmp/pgdata', { recursive: true })
